# backend/docs/swagger.yaml
definitions:
...
  internal_handler.CreateCommentRequest:
    properties:
      content:
        maxLength: 2000
        type: string
      parent_id:
        type: string
    required:
    - content
    type: object
  internal_handler.CreateEventRequest:
    properties:
      address:
        type: string
      cover_url:
        type: string
      description:
        type: string
      end_time:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      start_time:
        type: string
    required:
    - address
    - end_time
    - name
    - start_time
    type: object
  internal_handler.CreateOrderRequest:
    properties:
      idempotency_key:
        type: string
      items:
        items:
          properties:
            product_id:
              type: string
            quantity:
              type: integer
          required:
          - product_id
          - quantity
          type: object
        minItems: 1
        type: array
    required:
    - items
    type: object
  internal_handler.CreatePostRequest:
    properties:
      content:
        maxLength: 10000
        type: string
      images:
        items:
          type: string
        maxItems: 20
        type: array
      tags:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        maxLength: 200
        type: string
      video_url:
        type: string
    required:
    - content
    type: object
...
  internal_handler.PayRequest:
    properties:
      payment_id:
        type: string
      payment_method:
        enum:
        - wechat
        - alipay
        type: string
    required:
    - payment_id
    - payment_method
    type: object
...
  internal_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /events:
    get:
      description: List events with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event (requires auth)
      parameters:
      - description: Event info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - events
  /events/{id}:
    get:
      description: Get a single event by ID
      parameters:
      - description: Event UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Get event detail
      tags:
      - events
  /orders:
    get:
      description: List orders for the authenticated user (requires auth)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by order status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with product items (requires auth)
      parameters:
      - description: Order items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /orders/{order_no}:
    get:
      description: Get order details by order number (requires auth)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Get order detail
      tags:
      - orders
  /orders/{order_no}/cancel:
    post:
      description: Cancel a pending order (requires auth)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - orders
  /orders/{order_no}/pay:
    post:
      consumes:
      - application/json
      description: Record payment for an order (requires auth)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      - description: Payment method and ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Pay for an order
      tags:
      - orders
  /orders/{order_no}/refund:
    post:
      description: Process a refund for an order (requires auth)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Refund an order
      tags:
      - orders
...
  /posts:
    get:
      description: List posts with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post with content, images, and tags (requires auth)
      parameters:
      - description: Post info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - posts
  /posts/{id}:
    get:
      description: Get a single post by ID
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Get post detail
      tags:
      - posts
  /posts/{id}/comments:
    get:
      description: Get comments for a post
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: List post comments
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Add a comment to a post (requires auth)
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      - description: Comment content and optional parent_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - posts
...
  /products/{id}:
    get:
      description: Get a single product by ID
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Get product detail
      tags:
      - products
...
swagger: "2.0"